+++ app/src/main/java/com/battlelancer/seriesguide/movies/similar/SimilarMoviesFragment.kt
@@ -1,5 +1,6 @@
-package com.battlelancer.seriesguide.shows.search.similar
+package com.battlelancer.seriesguide.movies.similar
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.Menu
@@ -10,40 +11,49 @@ import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.MenuProvider
 import androidx.core.view.isGone
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.viewmodel.MutableCreationExtras
 import androidx.recyclerview.widget.RecyclerView
 import com.battlelancer.seriesguide.R
-import com.battlelancer.seriesguide.shows.search.discover.BaseAddShowsFragment
-import com.battlelancer.seriesguide.shows.search.discover.SearchResult
+import com.battlelancer.seriesguide.movies.search.MoviesSearchActivity
 import com.battlelancer.seriesguide.ui.AutoGridLayoutManager
-import com.battlelancer.seriesguide.ui.SearchActivity
 import com.battlelancer.seriesguide.ui.widgets.EmptyView
 import com.battlelancer.seriesguide.util.ThemeUtils
 import com.battlelancer.seriesguide.util.ViewTools
-import com.uwetrottmann.seriesguide.common.SingleLiveEvent
 import com.uwetrottmann.seriesguide.widgets.EmptyViewSwipeRefreshLayout
 
-class SimilarShowsFragment : BaseAddShowsFragment() {
+class SimilarMoviesFragment : Fragment() {
 
-    private var showTmdbId: Int = 0
-    private var showTitle: String? = null
+    private var tmdbId: Int = 0
+    private var title: String? = null
 
-    private val similarShowsViewModel: SimilarShowsViewModel by viewModels {
-        SimilarShowsViewModelFactory(requireActivity().application, showTmdbId)
-    }
+    private val viewModel: SimilarMoviesViewModel by viewModels(
+        extrasProducer = {
+            MutableCreationExtras(defaultViewModelCreationExtras).apply {
+                set(
+                    SimilarMoviesViewModel.KEY_TMDB_ID_MOVIE,
+                    requireArguments().getInt(ARG_TMDB_ID)
+                )
+            }
+        },
+        factoryProducer = { SimilarMoviesViewModel.Factory }
+    )
 
     private lateinit var swipeRefreshLayout: EmptyViewSwipeRefreshLayout
     private lateinit var emptyView: EmptyView
     private lateinit var recyclerView: RecyclerView
 
-    private lateinit var adapter: SimilarShowsAdapter
+    private lateinit var adapter: SimilarMoviesAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        showTmdbId = requireArguments().getInt(ARG_SHOW_TMDB_ID)
-        showTitle = requireArguments().getString(ARG_SHOW_TITLE)
+        tmdbId = requireArguments().getInt(ARG_TMDB_ID)
+        title = requireArguments().getString(ARG_TITLE)
+
+        (activity as AppCompatActivity).supportActionBar?.subtitle = title
     }
 
     override fun onCreateView(
@@ -51,7 +61,6 @@ class SimilarShowsFragment : BaseAddShowsFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Note: this layout is currently re-used by SimilarMoviesFragment.
         val v = inflater.inflate(R.layout.fragment_shows_similar, container, false)
         swipeRefreshLayout = v.findViewById(R.id.swipeRefreshLayoutShowsSimilar)
         emptyView = v.findViewById(R.id.emptyViewShowsSimilar)
@@ -65,11 +74,11 @@ class SimilarShowsFragment : BaseAddShowsFragment() {
 
         swipeRefreshLayout.apply {
             ViewTools.setSwipeRefreshLayoutColors(requireActivity().theme, this)
-            setOnRefreshListener { loadSimilarShows() }
+            setOnRefreshListener { load() }
         }
         emptyView.setButtonClickListener {
             swipeRefreshLayout.isRefreshing = true
-            loadSimilarShows()
+            load()
         }
 
         swipeRefreshLayout.isRefreshing = true
@@ -80,16 +89,16 @@ class SimilarShowsFragment : BaseAddShowsFragment() {
             layoutManager =
                 AutoGridLayoutManager(
                     context,
-                    R.dimen.showgrid_columnWidth,
+                    R.dimen.movie_grid_columnWidth,
                     1,
                     1
                 )
         }
 
-        adapter = SimilarShowsAdapter(itemClickListener)
+        adapter = SimilarMoviesAdapter(requireContext())
         recyclerView.adapter = adapter
 
-        similarShowsViewModel.resultLiveData.observe(viewLifecycleOwner) {
+        viewModel.resultLiveData.observe(viewLifecycleOwner) {
             adapter.submitList(it.results)
             emptyView.setMessage(it.emptyMessage)
             recyclerView.isGone = it.results.isNullOrEmpty()
@@ -104,14 +113,8 @@ class SimilarShowsFragment : BaseAddShowsFragment() {
         )
     }
 
-    private fun loadSimilarShows() {
-        similarShowsViewModel.loadSimilarShows(showTmdbId)
-    }
-
-    override fun onStart() {
-        super.onStart()
-        // Set (or restore if going back) the show title as action bar subtitle.
-        (activity as AppCompatActivity).supportActionBar?.subtitle = showTitle
+    private fun load() {
+        viewModel.loadSimilarMovies(tmdbId)
     }
 
     private val optionsMenuProvider = object : MenuProvider {
@@ -125,41 +128,27 @@ class SimilarShowsFragment : BaseAddShowsFragment() {
         override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
             return when (menuItem.itemId) {
                 MENU_ITEM_SEARCH_ID -> {
-                    startActivity(
-                        SearchActivity.newIntent(
-                            requireContext()
-                        )
-                    )
+                    startActivity(Intent(requireContext(), MoviesSearchActivity::class.java))
                     true
                 }
+
                 else -> false
             }
         }
     }
 
-    override fun setAllPendingNotAdded() {
-        similarShowsViewModel.setAllPendingNotAdded()
-    }
-
-    override fun setStateForTmdbId(showTmdbId: Int, newState: Int) {
-        similarShowsViewModel.setStateForTmdbId(showTmdbId, newState)
-    }
-
     companion object {
         val liftOnScrollTargetViewId = R.id.recyclerViewShowsSimilar
 
-        private const val ARG_SHOW_TMDB_ID = "ARG_SHOW_TMDB_ID"
-        private const val ARG_SHOW_TITLE = "ARG_SHOW_TITLE"
+        private const val ARG_TMDB_ID = "ARG_TMDB_ID"
+        private const val ARG_TITLE = "ARG_TITLE"
         private const val MENU_ITEM_SEARCH_ID = 1
 
-        @JvmStatic
-        val displaySimilarShowsEventLiveData = SingleLiveEvent<SearchResult>()
-
-        fun newInstance(showTmdbId: Int, showTitle: String?): SimilarShowsFragment {
-            return SimilarShowsFragment().apply {
+        fun newInstance(tmdbId: Int, title: String?): SimilarMoviesFragment {
+            return SimilarMoviesFragment().apply {
                 arguments = Bundle().apply {
-                    putInt(ARG_SHOW_TMDB_ID, showTmdbId)
-                    putString(ARG_SHOW_TITLE, showTitle)
+                    putInt(ARG_TMDB_ID, tmdbId)
+                    putString(ARG_TITLE, title)
                 }
             }
         }