+++ app/src/main/java/com/battlelancer/seriesguide/shows/ShowsFragment.kt
@@ -23,7 +23,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import androidx.lifecycle.whenStateAtLeast
+import androidx.lifecycle.withStarted
 import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -139,11 +139,11 @@ class ShowsFragment : Fragment() {
             // release time has passed).
             scheduledUpdateJob?.cancel()
             scheduledUpdateJob = viewLifecycleOwner.lifecycleScope.launch {
+                Timber.d("Scheduled query update.")
+                delay(DateUtils.MINUTE_IN_MILLIS + Random.nextLong(DateUtils.SECOND_IN_MILLIS))
                 // Use STARTED as in multi-window this might be visible, but not RESUMED.
                 // On the downside this runs even if the tab is not visible (tied to RESUMED).
-                viewLifecycleOwner.lifecycle.whenStateAtLeast(Lifecycle.State.STARTED) {
-                    Timber.d("Scheduled query update")
-                    delay(DateUtils.MINUTE_IN_MILLIS + Random.nextLong(DateUtils.SECOND_IN_MILLIS))
+                viewLifecycleOwner.lifecycle.withStarted {
                     updateShowsQuery()
                 }
             }
@@ -333,6 +333,7 @@ class ShowsFragment : Fragment() {
         override fun onSignInClicked() {
             startActivity(Intent(requireActivity(), MoreOptionsActivity::class.java))
             // Launching a top activity, adjust animation to match.
+            @Suppress("DEPRECATION") // just deprecated for predictive back
             requireActivity().overridePendingTransition(
                 R.anim.activity_fade_enter_sg, R.anim.activity_fade_exit_sg
             )
@@ -368,7 +369,7 @@ class ShowsFragment : Fragment() {
     }
 
     private val onPreferenceChangeListener =
-        OnSharedPreferenceChangeListener { _: SharedPreferences?, key: String ->
+        OnSharedPreferenceChangeListener { _: SharedPreferences?, key: String? ->
             if (key == AdvancedSettings.KEY_UPCOMING_LIMIT) {
                 updateShowsQuery()
                 // refresh all list widgets